+++ apps/decodex/src/agent/tracker_tool_bridge/tools/manual_attention/execution_event.rs
@@ -1,0 +1,63 @@
+use crate::{
+	agent::tracker_tool_bridge::{
+		self, ReviewHandoffContext, TrackerToolBridge,
+		tools::{
+			COMMENT_KIND_MANUAL_ATTENTION, MANUAL_ATTENTION_TERMINAL_PATH,
+			manual_attention::NormalizedManualAttentionComment,
+		},
+	},
+	tracker::records::{self, LinearExecutionEventIdentity, LinearExecutionEventRecord},
+};
+
+impl<'a> TrackerToolBridge<'a> {
+	pub(super) fn manual_attention_execution_event(
+		&self,
+		review_context: &ReviewHandoffContext,
+		comment: &NormalizedManualAttentionComment,
+	) -> LinearExecutionEventRecord {
+		let decision_request_id = comment
+			.decision_request
+			.as_ref()
+			.map(|request| request.decision_request_id.as_str())
+			.unwrap_or_default();
+		let anchor = records::stable_event_anchor(&[
+			COMMENT_KIND_MANUAL_ATTENTION,
+			comment.error_class.as_str(),
+			comment.next_action.as_str(),
+			comment.failed_command.as_deref().unwrap_or_default(),
+			comment.raw_error.as_deref().unwrap_or_default(),
+			decision_request_id,
+		]);
+		let mut record = LinearExecutionEventRecord::new(
+			LinearExecutionEventIdentity {
+				service_id: &review_context.service_id,
+				issue_id: &self.issue.id,
+				issue_identifier: &self.issue.identifier,
+				run_id: &review_context.run_id,
+				attempt_number: review_context.attempt_number,
+			},
+			"needs_attention",
+			tracker_tool_bridge::current_timestamp(),
+			&anchor,
+		);
+
+		record.branch = Some(review_context.branch_name.clone());
+		record.worktree_path = Some(review_context.worktree_path.clone());
+		record.pr_url = review_context.recorded_pr_url.clone();
+		record.summary = Some(
+			comment
+				.summary
+				.clone()
+				.unwrap_or_else(|| format!("Manual attention required: {}.", comment.error_class)),
+		);
+		record.error_class = Some(comment.error_class.clone());
+		record.next_action = Some(comment.next_action.clone());
+		record.blockers = Some(comment.blockers.clone());
+		record.evidence = Some(comment.evidence.clone());
+		record.terminal_path = Some(String::from(MANUAL_ATTENTION_TERMINAL_PATH));
+		record.failed_command = comment.failed_command.clone();
+		record.raw_error = comment.raw_error.clone();
+
+		record
+	}
+}
